Slope-grafiikat korjattu

+++ src/org/turkudragons/SpaceHunter/MapGame.java
@@ -110,8 +110,7 @@ public class MapGame extends BasicGameState {
 
 	@Override
 	public void render(GameContainer gc, StateBasedGame stg, Graphics g) throws SlickException {
-		g.translate(-(player.getX()-400),-(player.getY()-300));
-		
+		g.translate(-(int)(player.getX()-400),-(int)(player.getY()-300));
 		m.display();
 		//player.display(g);
 		for (Object o : oList) {
